Fully working support for service profiles
This breaks backward compatibility by removing the
ProxyMode setting in favor of a setting to control the
whole template that is used to render the frontend and
backend sections of the HAproxy config.

Also changes TemplateFile configuration in favor
of TemplateDir, with the new assumption that your main
file will be called haproxy.cfg.

Additional examples were added to the views/haproxy dir
to demonstrate how this is used.